another bug
1 + - x 666 = 0
This is because every operators + - x / will do the calculation if the tuple is valid.

this has been fixed by add a tracking variable isLastKeyInANumber to make sure operators won't do the calculation unless the previous key in is a number. If the previous key in is an operator, then just update tuple to overwrite the operator!

now
1 + - x 666 = 666